requirements.txt
asgiref==3.5.0
Django==3.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-parler==2.3
django-parler-rest==2.1
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
